requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
pillow
xgboost